Split transforms into Common, CPU and GPU parts and by strategy name.
