+++ compiler/src/iree/compiler/Codegen/LLVMCPU/BUILD
@@ -41,7 +41,7 @@ iree_compiler_cc_library(
         "//compiler/src/iree/compiler/Codegen/Dialect:IREECodegenDialect",
         "//compiler/src/iree/compiler/Codegen/Interfaces:PartitionableLoopsInterface",
         "//compiler/src/iree/compiler/Codegen/Sandbox",
-        "//compiler/src/iree/compiler/Codegen/TransformDialectStrategies/LLVMCPU:TransformDialectStrategies",
+        "//compiler/src/iree/compiler/Codegen/TransformDialectStrategies/CPU:TransformDialectStrategies",
         "//compiler/src/iree/compiler/Codegen/Transforms",
         "//compiler/src/iree/compiler/Codegen/Utils",
         "//compiler/src/iree/compiler/Codegen/VMVX",
