+++ compiler/src/iree/compiler/Codegen/LLVMCPU/CMakeLists.txt
@@ -85,7 +85,7 @@ iree_cc_library(
     iree::compiler::Codegen::Interfaces::PartitionableLoopsInterface
     iree::compiler::Codegen::PassHeaders
     iree::compiler::Codegen::Sandbox
-    iree::compiler::Codegen::TransformDialectStrategies::LLVMCPU::TransformDialectStrategies
+    iree::compiler::Codegen::TransformDialectStrategies::CPU::TransformDialectStrategies
     iree::compiler::Codegen::Transforms
     iree::compiler::Codegen::Utils
     iree::compiler::Codegen::VMVX
