+++ compiler/src/iree/compiler/Codegen/LLVMGPU/CMakeLists.txt
@@ -91,7 +91,7 @@ iree_cc_library(
     iree::compiler::Codegen::Dialect::IREECodegenDialect
     iree::compiler::Codegen::LLVMGPU::TransformExtensions::LLVMGPUExtensions
     iree::compiler::Codegen::PassHeaders
-    iree::compiler::Codegen::TransformDialectStrategies::LLVMGPU::TransformDialectStrategies
+    iree::compiler::Codegen::TransformDialectStrategies::GPU::TransformDialectStrategies
     iree::compiler::Codegen::Transforms
     iree::compiler::Codegen::Utils
     iree::compiler::Dialect::Flow::IR
