+++ compiler/src/iree/compiler/Codegen/TransformDialectStrategies/CPU/CMakeLists.txt
@@ -1,6 +1,6 @@
 ################################################################################
 # Autogenerated by build_tools/bazel_to_cmake/bazel_to_cmake.py from           #
-# compiler/src/iree/compiler/Codegen/TransformDialectStrategies/LLVMCPU/BUILD  #
+# compiler/src/iree/compiler/Codegen/TransformDialectStrategies/CPU/BUILD      #
 #                                                                              #
 # Use iree_cmake_extra_content from iree/build_defs.oss.bzl to add arbitrary   #
 # CMake-only content.                                                          #
@@ -14,9 +14,9 @@ iree_cc_library(
   NAME
     TransformDialectStrategies
   HDRS
-    "TransformDialectStrategiesCPU.h"
+    "Common.h"
   SRCS
-    "TransformDialectStrategiesCPU.cpp"
+    "Common.cpp"
   DEPS
     IREEDialectsTransforms
     IREELinalgExtDialect
