+++ compiler/src/iree/compiler/Codegen/TransformDialectStrategies/GPU/AbstractReductionStrategy.h
@@ -1,0 +1,73 @@
+// Copyright 2022 The IREE Authors
+//
+// Licensed under the Apache License v2.0 with LLVM Exceptions.
+// See https://llvm.org/LICENSE.txt for license information.
+// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
+
+#ifndef IREE_COMPILER_CODEGEN_TRANSFORM_DIALECT_STRATEGIES_GPU_ABSTRACT_REDUCTION_STRATEGY_H_
+#define IREE_COMPILER_CODEGEN_TRANSFORM_DIALECT_STRATEGIES_GPU_ABSTRACT_REDUCTION_STRATEGY_H_
+
+#include "iree-dialects/Transforms/TransformMatchers.h"
+#include "mlir/Dialect/GPU/IR/GPUDialect.h"
+
+namespace mlir {
+namespace iree_compiler {
+namespace gpu {
+
+/// Structure to hold a summary of HW-derived properties to configure the
+/// reduction strategy.
+/// The objective of this struct is to act as a minimal summary of key
+/// properties derived from the hardware (e.g. by an oracle) and that are
+/// sufficient to steer the strategy to produce a good version.
+/// These can be thought of as latent variables or embeddings that directly
+/// control the strategy and can be derived from the hardware by some procedure.
+struct ReductionConfig {
+  int64_t maxNumThreads;
+  int64_t vectorSize;
+};
+
+/// Structure to hold the parameters that control the reduction strategy.
+struct AbstractReductionStrategy {
+  AbstractReductionStrategy(
+      MLIRContext *context,
+      const transform_ext::MatchedReductionCaptures &captures)
+      : context(context), captures(captures) {
+    auto blockX =
+        mlir::gpu::GPUBlockMappingAttr::get(context, mlir::gpu::Blocks::DimX);
+    auto blockY =
+        mlir::gpu::GPUBlockMappingAttr::get(context, mlir::gpu::Blocks::DimY);
+    auto blockZ =
+        mlir::gpu::GPUBlockMappingAttr::get(context, mlir::gpu::Blocks::DimZ);
+    allBlockAttrs = SmallVector<Attribute>{blockX, blockY, blockZ};
+    auto threadX =
+        mlir::gpu::GPUThreadMappingAttr::get(context, mlir::gpu::Threads::DimX);
+    auto threadY =
+        mlir::gpu::GPUThreadMappingAttr::get(context, mlir::gpu::Threads::DimY);
+    auto threadZ =
+        mlir::gpu::GPUThreadMappingAttr::get(context, mlir::gpu::Threads::DimZ);
+    allThreadAttrs = SmallVector<Attribute>{threadX, threadY, threadZ};
+  }
+
+  virtual ~AbstractReductionStrategy() {}
+
+  virtual bool isProfitable() = 0;
+  virtual std::array<int64_t, 3> getNumThreadsInBlock() const = 0;
+
+  /// Constructor quantities.
+  MLIRContext *context;
+  transform_ext::MatchedReductionCaptures captures;
+
+  /// Derived quantities.
+  SmallVector<Attribute> allBlockAttrs;
+  SmallVector<Attribute> allThreadAttrs;
+
+  /// Tile sizes for the workgroup / determines grid size for all known
+  /// reduction strategies.
+  SmallVector<int64_t> workgroupTileSizes;
+};
+
+}  // namespace gpu
+}  // namespace iree_compiler
+}  // namespace mlir
+
+#endif  // IREE_COMPILER_CODEGEN_TRANSFORM_DIALECT_STRATEGIES_GPU_ABSTRACT_REDUCTION_STRATEGY_H_
