+++ compiler/src/iree/compiler/Codegen/TransformDialectStrategies/GPU/CMakeLists.txt
@@ -1,6 +1,6 @@
 ################################################################################
 # Autogenerated by build_tools/bazel_to_cmake/bazel_to_cmake.py from           #
-# compiler/src/iree/compiler/Codegen/TransformDialectStrategies/LLVMGPU/BUILD  #
+# compiler/src/iree/compiler/Codegen/TransformDialectStrategies/GPU/BUILD      #
 #                                                                              #
 # Use iree_cmake_extra_content from iree/build_defs.oss.bzl to add arbitrary   #
 # CMake-only content.                                                          #
@@ -14,9 +14,14 @@ iree_cc_library(
   NAME
     TransformDialectStrategies
   HDRS
-    "TransformDialectStrategiesGPU.h"
+    "AbstractReductionStrategy.h"
+    "Common.h"
+    "SmallReductionStrategy.h"
+    "StagedReductionStrategy.h"
   SRCS
-    "TransformDialectStrategiesGPU.cpp"
+    "Common.cpp"
+    "SmallReductionStrategy.cpp"
+    "StagedReductionStrategy.cpp"
   DEPS
     IREEDialectsTransforms
     IREELinalgExtDialect
